Move player stat DAL interfaces next to their interactor
Keep new DAL protocols (PlayerWithStatGetter, PlayedGamesByPlayerGetter)
in shvatka.core.players.interfaces instead of core.interfaces.dal.player,
co-located with the interactor that uses them.
